requirements.txt
# Specific dependencies to this package
numpy>=1.18.1
matplotlib>=3.1.3
pandas>=1.1.3
scikit-learn>=1.0.0
progressbar2>=3.55.0

# Finally install this package in editable (-e) 
# mode so I can develop it 
# without reinstalling it after every change.
-e .